Restructure arroyo-sql-macro dependencies
